Move helper functions for parsing info from requests to separate file
