+++ lib/general_utils.js
@@ -80,18 +80,6 @@ async function convertUnitsForCosmo (voiConfigs, timeseriesDataArray, parameter)
   return data
 }
 
-function checkValidityOfQuantityIds (voiConfigs) {
-  const voiExists = []
-  _.forEach(voiConfigs, function (voiConfig) {
-    if (_.isNil(_.get(voiConfig, ['report', 'key']))) {
-      voiExists.push(false)
-    } else {
-      voiExists.push(true)
-    }
-  })
-  return voiExists
-}
-
 function getVoiConfigsAsArray (vois, voisConfigs) {
   const voiConfigs = []
   _.forEach(vois, function (voi) {
@@ -103,21 +91,6 @@ function getVoiConfigsAsArray (vois, voisConfigs) {
   return voiConfigs
 }
 
-function getStationIdFromUrlPath (path) {
-  const splitUrl = path.split('/')
-  const sid = splitUrl[2]
-  return sid
-}
-
-function getVoisNamesFromQuery (query) {
-  const voi = query.quantities
-  const defaultParameter = ['t_2m']
-  let vois = defaultParameter
-  if (voi) {
-    vois = voi.split(',')
-  }
-  return vois
-}
 
 function getCSVLabels (voiConfigs) {
   let csvLabels = 'timestamp / ms,'
@@ -190,13 +163,10 @@ function getStringsFromStationDataPayload (timeseriesDataArray) {
 }
 
 exports.getVoiConfigsAsArray = getVoiConfigsAsArray
-exports.getStationIdFromUrlPath = getStationIdFromUrlPath
-exports.getVoisNamesFromQuery = getVoisNamesFromQuery
 exports.getCSVLabels = getCSVLabels
 exports.getHeadElementsFromTimeseries = getHeadElementsFromTimeseries
 exports.getValuesAsCSVString = getValuesAsCSVString
 exports.getStringsFromStationDataPayload = getStringsFromStationDataPayload
-exports.checkValidityOfQuantityIds = checkValidityOfQuantityIds
 exports.convertUnitsForMosmix = convertUnitsForMosmix
 exports.convertUnitsForCosmo = convertUnitsForCosmo
 exports.getVoiConfigsAsArray = getVoiConfigsAsArray
