+++ lib/measured_values_utils.js
@@ -9,18 +9,6 @@ const gu = require('../lib/general_utils.js')
 
 // General functions ---------------------------------------------------------------- //
 
-function checkValidityOfQuantityIds (voiConfigs) {
-  const voiExists = []
-  _.forEach(voiConfigs, function (voiConfig) {
-    if (_.isNil(_.get(voiConfig, ['report', 'key']))) {
-      voiExists.push(false)
-    } else {
-      voiExists.push(true)
-    }
-  })
-  return voiExists
-}
-
 function dropTimeseriesDataNotOfInterest (voiConfigs, timeseriesDataCollection) {
   const keys = _.map(voiConfigs, function (o) {
     return o.report.key
@@ -174,7 +162,6 @@ function renderMeasuredValuesAsCSV (voiConfigs, timeseriesDataArray) {
 }
 
 // Export functions
-exports.checkValidityOfQuantityIds = checkValidityOfQuantityIds
 exports.renderMeasuredValuesAsCSV = renderMeasuredValuesAsCSV
 exports.renderMeasuredValuesAsJSON = renderMeasuredValuesAsJSON
 exports.dropTimeseriesDataNotOfInterest = dropTimeseriesDataNotOfInterest
